requirements.txt

flask-socketio==5.3.6
eventlet==0.33.1
dnspython==2.0.0
websockets
beautifulsoup4==4.10.0
openai==0.27.4
astor==0.8.1
tiktoken==0.5.1
numpy
pandas==1.4.4
Werkzeug
json5==0.9.6
urllib3==1.26.11
sortedcollections==2.1.0
sortedcontainers==2.4.0
matplotlib==3.5.2
typing-extensions==4.5.0
async-timeout==4.0.2
attrs==23.1.0
dnszone
bs4==0.0.1
cachelib==0.10.2
cachetools==4.2.4
document-utils==1.8.1
fastapi==0.99.1
Flask
flask
Flask-Cors==4.0.0
Flask-Session==0.5.0
fsspec
grpcio-status==1.48.2
gunicorn==21.2.0
html5lib==1.1
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.0
pip==23.3.2
wheel==0.42.0
setuptools==69.0.2
requests==2.31.0
plotly==5.18.0
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio
python-magic==0.4.27
requests-oauthlib==1.3.1
scikit-learn==1.3.0
scipy
semi==0.0.4
simple-websocket==1.0.0
SQLAlchemy==2.0.19
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
uvicorn==0.23.2
vectorhub==1.8.3
watchfiles==0.19.0
webencodings==0.5.1
collections-mapping
google-auth==2.25.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.0.0
gcsfs==2023.12.2.post1
Jinja2==3.1.2
google-cloud-storage==2.14.0
google-cloud==0.34.0
